Runnel: Perf improvement in sizing calculations.
When we encode, the first thing that happens is we ask what
the size of the encoded object will be, so we can create
the proper size buffer for output. This means a traversal
of the entire runnel struct, which can be quite nested.

But... then, as it traverses the structures, sub structures
do the same thing. So you get a tinyish NLogM  blow up
in counting sizes.

It's a small thing, but noticable.

By caching the valueSize() call in containing structures
(StructEncoder, ArrayEncoder) we only calculate the size once.

Additionally, I fixed 2 tests that were broken on Windows.