Validate DC object names to prevent the use of special words or characters.
This change is required to be able to control the user input for names that are mostly used at different key purposes like filenames, etc. DC expects the node name to be used inside the config filename, and DC is relying on that to detect which node name to start when a config dir is used. Also, DC config file format and CLI input to change the configuration rely on cluster, stripe and node names to be used as identifiers for the change, which prevents some special characters to be used.

Notes:

- This change will require an XPC
- This change will highly likely be compatible with existing DC users
- This change will highly likely be compatible with existing tc-config.xml: their node name will be sanitized and transformed to the same value used when persisting data

Fix made in this commit:

- It was not possible to migrate a tc-config.xml file having servers with no name (server names should be generated from `host:port`)