Remove POM repositories elements & bump terracotta-utilities version
This commit bumps the terracotta-parent (and other dependency) versions
to reduce the repositories and pluginRepositories element to those
required to locate direct dependencies of this project.  More
specifically, SNAPSHOT repositories are removed -- inclusion of
SNAPSHOT repositories interferes with the use of range-based version
specifications.  Should a developer want to include a SNAPSHOT release
in a build, local addition of the SNAPSHOT repositories can be employed.

This commit also bumps the version of terracotta-utilities used.