Improve user feedback for config-tool import CLI
This change allows the import command to validate if a node we are connected to is really part of the cluster topology we would like to import.

Importing a topology on a node that is not part of the imported cluster will do a no-op and the node won't be updated.

This change allows to fail the config-tool import command with an error message if the user tries to import a cluster topology in a node that has no reference inside this cluster topology.